requirements.txt
python-dotenv==1.0.0
telethon==1.30.1
pymsteams==0.2.2
cryptg==0.4.0
rarfile==4.1
httpx==0.25.1